Project/requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
distlib==0.3.7
Django==4.2.6
filelock==3.12.4
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
MarkupSafe==2.1.3
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
pipenv==2023.10.3
platformdirs==3.11.0
pydub==0.25.1
pygame==2.5.2
pyserial==3.5
requests==2.31.0
speedtest==0.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5